Separate the code
We need that to check whether two given hash bounds are equal into a
separate function, so that it can be called from multiple places.  Right
now it's only caller is partition_bounds_equal() but later we will use
it for merging partition bounds.

Ashutosh Bapat, reviewed by Dmitry Dolgov