Extra tests
The tests added in partition_join for advanced partition matching are
minimal.  We need to move some of the tests here in that file and
discard remaining tests. For now, keep it in separate file till we
resolve all the issues.

THIS FILE IS NOT INTENDED TO BE COMMITTED.

Rajkumar Raghuvanshi and Ashutosh Bapat.