Add mean stats
Just couple of them for the start, namely mean waiting time in pending
and mean pending depth. Those stats are not bounded directly to a
backed, since the idea to extend it as the follow up for different type
of rings (specifically for WAL/vacuum etc). In case of small number of
io those times/depths are usually neglegitable, so record only in case
of relatively big queues based on track_queue_threshold.