Track progress of undo execution.
The code relying on the undo log infrastructure (e.g.the zheap AM in the
future) may need to know whether particular undo record has been
applied. Besides that, the information on "undo execution progress" is
important for undo log discarding, which is implemented by the next patch of
the series.

Here we add a new field last_rec_applied to the undo record set chunk header
and provide the user two functions to update it:

   * UndoPrepareToUpdateLastAppliedRecord() - like other UndoPrepareTo*
     functions, this finds and locks buffers containing the last_rec_applied
     field.

   * UpdateLastAppliedRecord() performs the actual update and adds information
     to WAL such that the last_rec_applied field is also updated during crash
     recovery.

The caller is responsible for setting LSN of affected buffers and for
unlocking them - convenience functions might need to be added for this.

Author: Antonin Houska