Const reduction
Current implementation does merge only simple cases when a query
contains series of constants. As soon as those constants are wrapped
into an expression or function it's not merged anymore. Improve this by
evaluating necessary expressions into constants first, then proceed with
the merge logic.

For that purpose we introduce a constant walker that return all found
constants in the expression. The trick is that simple evaluation
produces constants without a correct location and we're not able to
record it.  Avoid this by finding the actual constants and picking the
last one as the final location of the expression.