Use the undo log to clean up database directory.
If the server crashes while performing commands CREATE DATABASE or ALTER
DATABASE ... SET TABLESPACE before the catalog changes could be committed, the
database files can be left on disk. Similar to the creation of relation
storage, we ensure the cleanup by creating the appropriate undo record before
the new database directory is created. The undo is replayed right after crash
recovery and removes the orphaned files.

Author: Antonin Houska