WIP: Reinitialize buffer sync strategy
Resizing the shared buffers renders the saved state of BgBufferSync()
invalid.  Hence reinitialize it. The state is saved in static variables
inside the function and thus can not be accessed from outside the
function. Hence we add an argument to BgBufferSync() to request the
function to reset the state.

TODO: Ideally we should save this state in some global structure and add
a function to reset it.

TODO: StrategyInitialize() initializes buffer lookup table but
StrategyReInitialize() doesn't.  Where should we reinitialize the buffer
lookup table?